requirements.txt
pyFneko>=3.2
